requirements.txt
python-dotenv==1.0.1
haystack-ai==2.2.4
haystack-experimental==0.1.0
sentence-transformers==3.0.1 
cohere-haystack==1.1.3
colorama==0.4.6
trafilatura==1.11.0
geopy==2.2.0
pandas==2.2.2